Rework and improve the help output
Also use Stdout for all our outputs; we'll leave the stderr for the
Golang runtime, since I plan to scrape it with another process and
use that for error reporting, maybe something like panicwrap or
panicwatch, ensuring that we don't interfere with systemd support
or signals.

Also, using help now shows the home page for the project and it
no longer exits with an error code.

Signed-off-by: Jeffrey H. Johnson <trnsz@pobox.com>